refactor(verify-step2): scaffold split behind stable facade (#349)
* refactor(verify-step2): add compile-safe split scaffold

* refactor(verify-step2): mechanically move verify logic behind facade

* docs(verify-step2): add move-map and boundary reviewer gates

* ci/docs(verify-step2): tighten boundary gates and test-safe leakage checks

* ci/docs(wave5-step2): add verifyresult and canonicalization single-source gates
