+++ crates/assay-registry/src/verify_next/errors.rs
@@ -1,0 +1,5 @@
+//! Error mapping boundary for verify split.
+//!
+//! Contract target:
+//! - shared error-construction helpers only
+//! - no ownership of parsing/crypto logic
