+++ crates/assay-registry/src/verify_next/tests.rs
@@ -1,0 +1,4 @@
+#![cfg(test)]
+
+//! Placeholder test module for verify split scaffold.
+//! No test moves in Commit B; Step1 contract tests remain in `verify.rs`.
