+++ docs/architecture/PLAN-split-refactor-2026q1.md
@@ -87,7 +87,8 @@ Step status:
 
 - Writer split: merged via PR #332.
 - Verify split: pending completion.
-- Verify Step1 behavior freeze (tests/docs/gates): in progress on `codex/wave5-step1-verify-freeze`.
+- Verify Step1 behavior freeze (tests/docs/gates): in review on `codex/wave5-step1-verify-freeze` (PR #348).
+- Verify Step2 mechanical split: in progress on `codex/wave5-step2-verify-split` (PR #349).
 
 ### A. `crates/assay-registry/src/verify.rs`
